+++ src/shared/particle_dynamics/solid_dynamics/contact_dynamics/all_contact_dynamics.h
@@ -21,7 +21,7 @@
  *                                                                           *
  * ------------------------------------------------------------------------- */
 /**
- * @file    all_solid_dynamics.h
+ * @file    all_contact_dynamics.h
  * @brief   This is the header file that user code should include to pick up all
  *          solid dynamics used in SPHinXsys.
  * @author	Chi Zhang and Xiangyu Hu
@@ -29,12 +29,6 @@
 
 #pragma once
 
-#include "all_contact_dynamics.h"
-#include "constraint_dynamics.h"
-#include "elastic_dynamics.h"
-#include "fluid_structure_interaction.h"
-#include "general_solid_dynamics.h"
-#include "inelastic_dynamics.h"
-#include "loading_dynamics.h"
-#include "thin_structure_dynamics.h"
-#include "thin_structure_math.h"
+#include "contact_friction.h"
+#include "contact_repulsion.h"
+#include "repulsion_density_summation.h"
