+++ src/shared/particle_dynamics/solid_dynamics/contact_dynamics/base_contact_dynamics.h
@@ -21,20 +21,24 @@
  *                                                                           *
  * ------------------------------------------------------------------------- */
 /**
- * @file    all_solid_dynamics.h
- * @brief   This is the header file that user code should include to pick up all
- *          solid dynamics used in SPHinXsys.
+ * @file 	base_contact_dynamics.h
+ * @brief 	Here, we define the algorithm classes for solid contact dynamics.
+ * @details We consider here a weakly compressible solids.
  * @author	Chi Zhang and Xiangyu Hu
  */
 
-#pragma once
+#ifndef BASE_CONTACT_DYNAMICS_H
+#define BASE_CONTACT_DYNAMICS_H
 
-#include "all_contact_dynamics.h"
-#include "constraint_dynamics.h"
-#include "elastic_dynamics.h"
-#include "fluid_structure_interaction.h"
 #include "general_solid_dynamics.h"
-#include "inelastic_dynamics.h"
-#include "loading_dynamics.h"
-#include "thin_structure_dynamics.h"
-#include "thin_structure_math.h"
+
+namespace SPH
+{
+namespace solid_dynamics
+{
+typedef DataDelegateContact<SolidParticles, SolidParticles> ContactDynamicsData;
+typedef DataDelegateContact<SolidParticles, SolidParticles> ContactWithWallData;
+
+} // namespace solid_dynamics
+} // namespace SPH
+#endif // BASE_CONTACT_DYNAMICS_H
