+++ tests/optimization/test_2d_VP_same_sink_optimization/VP_same_sink_optimization.cpp
@@ -491,7 +491,7 @@ TEST(test_optimization, test_problem1_optimized)
     tt = t2 - t1;
     std::cout << "Total time for optimization: " << tt.seconds() << " seconds." << std::endl;
 
-    EXPECT_GT(500.0, calculate_averaged_opt_temperature.exec());
+    EXPECT_GT(502.0, calculate_averaged_opt_temperature.exec());
 };
 
 int main(int argc, char *argv[])
